+++ src/org/schema/constants/traits/DeliveryEvent.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace org\schema\constants\traits;
+
+/**
+ * The DeliveryEvent properties enumeration.
+ */
+trait DeliveryEvent
+{
+    const string ACCESS_CODE         = 'accessCode'        ;
+    const string AVAILABLE_FROM      = 'availableFrom'     ;
+    const string AVAILABLE_THROUGH   = 'availableThrough'  ;
+    const string HAS_DELIVERY_METHOD = 'hasDeliveryMethod' ;
+}