+++ src/org/schema/constants/traits/Properties.php
@@ -29,6 +29,7 @@ trait Properties
         DefinedTerm ,
         DefinedTermSet ,
         DeliveryChargeSpecification ,
+        DeliveryEvent ,
         Demand ,
         DublinCore ,
         EndorsementRating ,